a bit of a mess, but has many parts started
