+++ stocks_api/models/.weather_location.py
